+++ application/models/Site_Model.php
@@ -902,7 +902,7 @@ class MangaCow extends Site_Model {
 
 class SeaOtterScans extends Site_Model {
 	public function getFullTitleURL(string $title_url) : string {
-		return "http://reader.seaotterscans.com/series/{$title_url}";
+		return "https://reader.seaotterscans.com/series/{$title_url}";
 	}
 
 	public function isValidTitleURL(string $title_url) : bool {
@@ -959,7 +959,7 @@ class SeaOtterScans extends Site_Model {
 
 				$link = (string) $nodes_chapter->item(0)->getAttribute('href');
 				$titleData['latest_chapter'] = preg_replace('/.*\/read\/.*?\/(.*?)\/$/', '$1', $link);
-				$titleData['last_updated'] =  date("Y-m-d H:i:s", strtotime((string) str_replace('.', '', explode(',', $nodes_latest[0]->textContent)[1])));
+				$titleData['last_updated'] = (string) str_replace('.', '', explode(',', $nodes_latest[0]->textContent)[1]);
 			}
 		} else {
 			//TODO: Throw ERRORS;